+++ packages/@expo/cli/src/export/createBundles.ts
@@ -17,6 +17,9 @@ export async function createBundlesAsync(
   publishOptions: PublishOptions = {},
   bundleOptions: { platforms: Platform[]; dev?: boolean }
 ): Promise<Partial<Record<Platform, BundleOutput>>> {
+  if (!bundleOptions.platforms.length) {
+    return {};
+  }
   const projectConfig = getConfig(projectRoot, { skipSDKVersionRequirement: true });
   const { exp } = projectConfig;
 
@@ -41,6 +44,7 @@ export async function createBundlesAsync(
       platform,
       entryPoint: getEntryWithServerRoot(projectRoot, projectConfig, platform),
       dev: bundleOptions.dev,
+      sourceMapUrl: undefined,
     }))
   );