+++ packages/expo-modules-core/ios/Swift/SharedObjects/SharedObjectRegistry.swift
@@ -126,4 +126,8 @@ public final class SharedObjectRegistry {
     }
     return createSharedJavaScriptObject(runtime: runtime, nativeObject: nativeObject)
   }
+
+  internal static func clear() {
+    pairs.removeAll()
+  }
 }