+++ packages/expo-updates/android/src/main/java/expo/modules/updates/loader/FileDownloader.kt
@@ -93,6 +93,38 @@ open class FileDownloader(context: Context, private val client: OkHttpClient) {
   }
 
   internal fun parseRemoteUpdateResponse(response: Response, configuration: UpdatesConfiguration, callback: RemoteUpdateDownloadCallback) {
+    val responseHeaders = response.headers
+    val responseHeaderData = ResponseHeaderData(
+      protocolVersionRaw = responseHeaders["expo-protocol-version"],
+      manifestFiltersRaw = responseHeaders["expo-manifest-filters"],
+      serverDefinedHeadersRaw = responseHeaders["expo-server-defined-headers"],
+      manifestSignature = responseHeaders["expo-manifest-signature"],
+    )
+    val responseBody = response.body
+
+    if (response.code == 204 || responseBody == null) {
+      // If the protocol version greater than 0, we support returning a 204 and no body to mean no-op.
+      // A 204 has no content-type.
+      if (responseHeaderData.protocolVersion != null && responseHeaderData.protocolVersion > 0) {
+        callback.onSuccess(
+          UpdateResponse(
+            responseHeaderData = responseHeaderData,
+            manifestUpdateResponsePart = null,
+            directiveUpdateResponsePart = null
+          )
+        )
+        return
+      }
+
+      val message = "Missing body in remote update"
+      logger.error(message, UpdatesErrorCode.UpdateFailedToLoad)
+      callback.onFailure(
+        message,
+        IOException(message)
+      )
+      return
+    }
+
     val contentType = response.header("content-type") ?: ""
     val isMultipart = contentType.startsWith("multipart/", ignoreCase = true)
     if (isMultipart) {
@@ -107,17 +139,8 @@ open class FileDownloader(context: Context, private val client: OkHttpClient) {
         return
       }
 
-      parseMultipartRemoteUpdateResponse(response, boundaryParameter, configuration, callback)
+      parseMultipartRemoteUpdateResponse(responseBody, responseHeaderData, boundaryParameter, configuration, callback)
     } else {
-      val responseHeaders = response.headers
-
-      val responseHeaderData = ResponseHeaderData(
-        protocolVersion = responseHeaders["expo-protocol-version"],
-        manifestFiltersRaw = responseHeaders["expo-manifest-filters"],
-        serverDefinedHeadersRaw = responseHeaders["expo-server-defined-headers"],
-        manifestSignature = responseHeaders["expo-manifest-signature"]
-      )
-
       val manifestResponseInfo = ResponsePartInfo(
         responseHeaderData = responseHeaderData,
         responsePartHeaderData = ResponsePartHeaderData(
@@ -165,13 +188,13 @@ open class FileDownloader(context: Context, private val client: OkHttpClient) {
     return headers.toHeaders()
   }
 
-  private fun parseMultipartRemoteUpdateResponse(response: Response, boundary: String, configuration: UpdatesConfiguration, callback: RemoteUpdateDownloadCallback) {
+  private fun parseMultipartRemoteUpdateResponse(responseBody: ResponseBody, responseHeaderData: ResponseHeaderData, boundary: String, configuration: UpdatesConfiguration, callback: RemoteUpdateDownloadCallback) {
     var manifestPartBodyAndHeaders: Pair<String, Headers>? = null
     var extensionsBody: String? = null
     var certificateChainString: String? = null
     var directivePartBodyAndHeaders: Pair<String, Headers>? = null
 
-    val multipartStream = MultipartStream(response.body!!.byteStream(), boundary.toByteArray())
+    val multipartStream = MultipartStream(responseBody.byteStream(), boundary.toByteArray())
 
     try {
       var nextPart = multipartStream.skipPreamble()
@@ -226,14 +249,6 @@ open class FileDownloader(context: Context, private val client: OkHttpClient) {
       callback.onFailure(message, IOException(message))
       return
     }
-
-    val responseHeaders = response.headers
-    val responseHeaderData = ResponseHeaderData(
-      protocolVersion = responseHeaders["expo-protocol-version"],
-      manifestFiltersRaw = responseHeaders["expo-manifest-filters"],
-      serverDefinedHeadersRaw = responseHeaders["expo-server-defined-headers"],
-      manifestSignature = responseHeaders["expo-manifest-signature"],
-    )
 
     val manifestResponseInfo = manifestPartBodyAndHeaders?.let {
       ResponsePartInfo(